// src/enums/colorEnum.ts
export enum StatColor {
    offense = "#E63946",
    defense = "#6bd55d",
    technique = "#ff9728",
    speed = "#6fd5e7",
    spirit = "#b75ae0",
    dice = "grey"
}

export enum PlayerColor {
    Player = "#3B82F6",
    Enemy = "#dc5555",
}

export enum ElementColor {
    Normal = "#A8A77A",
    Fire = "#EE8130",
    Water = "#6390F0",
    Electric = "#F7D02C",
    Grass = "#7AC74C",
    Ice = "#96D9D6",
    Fighting = "#C22E28",
    Poison = "#A33EA1",
    Ground = "#E2BF65",
    Flying = "#A98FF3",
    Psychic = "#F95587",
    Bug = "#A6B91A",
    Rock = "#B6A136",
    Ghost = "#735797",
    Dragon = "#6F35FC",
    Dark = "#705746",
    Steel = "#B7B7CE",
    Fairy = "#D685AD",
}